Split Linkify file logic to the several files
